requirements.txt
dash==2.9.1
matplotlib==3.6.2
mysql_connector_repackaged==0.3.1
pandas==1.5.2
seaborn==0.12.2
gunicorn
dash-tools
